App/backend_drf/requirements.txt
asgiref==3.7.2
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
dj-database-url==2.1.0
Django==4.2.7
djangorestframework==3.14.0
environs==9.5.0
exceptiongroup==1.1.3
executing==0.8.3
fonttools==4.46.0
iniconfig==2.0.0
joblib==1.3.2
kiwisolver==1.4.5
marshmallow==3.20.1
mat2py==0.0.1
matplotlib==3.8.2
networkx==3.2
numpy==1.26.2
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
pluggy==1.3.0
ply==3.11
pyparsing==3.1.1
pytest==7.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
six==1.16.0
smop==0.41b0
sqlparse==0.4.4
threadpoolctl==3.2.0
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
